refactor: share button
