+++ lib/share/view/view.dart
@@ -1,5 +1,6 @@
 export 'facebook_button.dart';
 export 'share_bottom_sheet.dart';
+export 'share_button.dart';
 export 'share_dialog.dart';
 export 'share_page.dart';
 export 'twitter_button.dart';
